feat: refactor and new queries
