+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@frmscoe/frms-coe-lib",
-  "version": "0.2.1",
+  "version": "0.3.0",
   "description": "FRMS Center of Excellence package library",
   "main": "lib/index.js",
   "repository": {
@@ -22,7 +22,7 @@
   "dependencies": {
     "@log4js-node/logstash-http": "^1.1.0",
     "@types/uuid": "^9.0.2",
-    "arangojs": "^8.3.0",
+    "arangojs": "^8.4.0",
     "dotenv": "^16.3.1",
     "ioredis": "^5.3.2",
     "log4js": "^6.9.1",
@@ -45,12 +45,12 @@
     "eslint-plugin-prettier": "^4.2.1",
     "eslint-plugin-promise": "^6.1.1",
     "husky": "^8.0.3",
-    "ioredis-mock": "^8.7.0",
+    "ioredis-mock": "^8.8.1",
     "jest": "^29.5.0",
     "prettier": "^2.8.8",
     "ts-jest": "^29.1.0",
     "ts-node": "^10.9.1",
-    "tslib": "^2.6.0",
+    "tslib": "^2.6.1",
     "typescript": "^5.1.5"
   }
 }
