+++ src/index.ts
@@ -2,12 +2,12 @@ import { aql } from 'arangojs';
 import { type GeneratedAqlQuery } from 'arangojs/aql';
 import {
   CreateDatabaseManager,
-  type ConfigurationDB,
   type DatabaseManagerInstance,
   type ManagerConfig,
-  type PseudonymsDB,
-  type TransactionHistoryDB,
 } from './services/dbManager';
+import { type ConfigurationDB } from './interfaces/database/ConfigurationDB';
+import { type TransactionHistoryDB } from './interfaces/database/TransactionHistoryDB';
+import { type PseudonymsDB } from './interfaces/database/PseudonymsDB';
 import { LoggerService } from './services/logger';
 import { RedisService } from './services/redis';
